Update arpack-ng-config.cmake to match pkgconfig files
Removes the manual search for library files (where the static lib was
preferred by default), instead passes the library name as a -l flag so
the system can search for it in the preferred manner.

MPI libraries and includes should not be part of PARPACK's interface,
instead they belong in the equivalent of Libs.private/Requires.private
in the pkgconfig files.  Remove them.

PARPACK has no header files of its own.  Remove them from the
target_include_directories stuff.