[BUG FIX] Make sure iseed is always initialized to values allowed by lapack ?larnv.
This means killing inits variable:
- Using MPI, inits (shared variable) may be set to false by one proc,
and prevent other procs to initialize seeds (as inits is shared by
use of the save fortran keyword).
- Not using MPI, inits only prevents from re-initializing seeds which
have already been initialized.

This commit is not-op. From a functional point of view, we are doing
the same thing. From an implementation point of view, we make sure
iseed is always (for all MPI procs) initialized to values allowed
by lapack (if not, lapack crashes).

This problem doesn't occur with sequential code. Changes have been done
in both sequential and MPI code to keep things symmetric.